pyproject.toml
[project]
name = "bench-backend"
version = "0.1.0"
description = "Decentralized community LLM benchmarking platform"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi==0.104.1",
    "sqlmodel==0.0.14",
    "uvicorn[standard]==0.24.0",
    "redis==5.0.1",
    "python-jose[cryptography]==3.3.0",
    "httpx==0.25.2",
    "alembic==1.13.1",
    "python-multipart==0.0.6",
    "typer==0.9.0",
    "psycopg2-binary==2.9.10",
    "pydantic-settings==2.9.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[project.scripts]
bench = "apps.bench_cli.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "A"]
ignore = ["E501"]  # Line too long

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true